Initial Upload
MagCrysAnisFromCIF is a module containing the main functions useed in the rest of the code. UVW and prop tensor main generates csv files for the uvw choices and property tensor components of sites on a crystal using a cif file. plotting energy surfaces plots energy density surfaces onto a 3d plot of a unit cell. The vesta file maker plots uvw arrows onto vesta file for a crystal.